+++ src/components/terms/jira/MathOperations.php
@@ -1,10 +1,13 @@
 <?php
 namespace extas\components\terms\jira;
 
+use extas\components\exceptions\MissedOrUnknown;
 use extas\components\Item;
 use extas\components\THasClass;
 use extas\interfaces\IHasClass;
 use extas\interfaces\terms\ITerm;
+use extas\interfaces\terms\jira\results\ICalculationResult;
+use extas\interfaces\terms\jira\results\IResultIssues;
 use extas\interfaces\terms\jira\strategies\IMathOperationStrategy;
 
 /**
@@ -43,13 +46,15 @@ class MathOperations extends JiraTermCalculator
     public const TERM_PARAM__STRATEGY = 'strategy';
 
     protected string $marker = self::TERM_PARAM__MARKER;
+    protected string $argsInterface = IResultIssues::class;
 
     /**
      * @param ITerm $term
-     * @param array $args
-     * @return mixed|null
+     * @param array $issues
+     * @return ICalculationResult
+     * @throws MissedOrUnknown
      */
-    protected function execute(ITerm $term, array $args)
+    protected function execute(ITerm $term, array $issues): ICalculationResult
     {
         /**
          * @var IMathOperationStrategy $strategy
@@ -58,13 +63,13 @@ class MathOperations extends JiraTermCalculator
         $strategyClass = $term->getParameterValue(static::TERM_PARAM__STRATEGY, '');
 
         if (!$strategyClass) {
-            return null;
+            throw new MissedOrUnknown('strategy class');
         }
 
         $withClass = $this->getStrategyHolder($strategyClass);
         $strategy = $withClass->buildClassWithParameters();
 
-        return $strategy($this, $term, $this->getIssues($args));
+        return $strategy($this, $term, $issues);
     }
 
     /**