+++ src/components/terms/jira/TotalByField.php
@@ -1,7 +1,10 @@
 <?php
 namespace extas\components\terms\jira;
 
+use extas\components\terms\jira\results\ResultNumeric;
 use extas\interfaces\terms\ITerm;
+use extas\interfaces\terms\jira\results\ICalculationResult;
+use extas\interfaces\terms\jira\results\IResultIssues;
 
 /**
  * Class TotalByField
@@ -12,24 +15,24 @@ use extas\interfaces\terms\ITerm;
 class TotalByField extends JiraTermCalculator
 {
     use THasIssuesSearchResult;
+    use THasFieldSubfield;
 
     public const TERM_PARAM__MARKER = 'jira__by_field_sum';
     public const TERM_PARAM__FIELD_NAME = 'field_name';
-    public const TERM_PARAM__SUB_FIELD_NAME = 'sub_field_name';
+    public const TERM_PARAM__SUB_FIELD_NAME = 'subfield_name';
 
     protected string $marker = self::TERM_PARAM__MARKER;
+    protected string $argsInterface = IResultIssues::class;
 
     /**
      * @param ITerm $term
-     * @param array $args
-     * @return int|mixed|null
+     * @param array $issues
+     * @return ICalculationResult
      */
-    protected function execute(ITerm $term, array $args)
+    protected function execute(ITerm $term, array $issues): ICalculationResult
     {
-        $fieldName = $term->getParameterValue(static::TERM_PARAM__FIELD_NAME, '');
-        $subfield = $term->getParameterValue(static::TERM_PARAM__SUB_FIELD_NAME, '');
-        $subfieldMethod = $subfield ? 'getField' . ucfirst($subfield) : 'getFieldValue';
-        $issues = $this->getIssues($args);
+        $fieldName = $this->getField($term);
+        $subfieldMethod = $this->getSubfieldMethod($term);
         $total = 0;
 
         foreach ($issues as $issue) {
@@ -39,6 +42,8 @@ class TotalByField extends JiraTermCalculator
             }
         }
 
-        return $total;
+        return new ResultNumeric([
+            ResultNumeric::FIELD__NUMBER => $total
+        ]);
     }
 }