+++ src/components/terms/jira/strategies/MathOperationTotal.php
@@ -2,10 +2,12 @@
 namespace extas\components\terms\jira\strategies;
 
 use extas\components\terms\jira\MathOperations;
+use extas\components\terms\jira\results\ResultNumeric;
 use extas\components\terms\jira\THasIssuesSearchResult;
 use extas\interfaces\jira\issues\IIssue;
 use extas\interfaces\terms\ITerm;
 use extas\interfaces\terms\ITermCalculator;
+use extas\interfaces\terms\jira\results\ICalculationResult;
 
 /**
  * Class MathOperationTotal
@@ -21,15 +23,13 @@ class MathOperationTotal extends MathOperationStrategy
     use THasIssuesSearchResult;
     use TMathOperation;
 
-    public const TERM_PARAM__TOTAL_OPERATION = 'operation_total';
-
     /**
      * @param ITermCalculator $calculator
      * @param ITerm $term
      * @param IIssue[] $issues
-     * @return mixed|void
+     * @return ICalculationResult
      */
-    public function __invoke(ITermCalculator $calculator, ITerm $term, array $issues)
+    public function __invoke(ITermCalculator $calculator, ITerm $term, array $issues): ICalculationResult
     {
         $operation = $term->getParameterValue(MathOperations::TERM_PARAM__OPERATION, '');
         $fieldsNames = $term->getParameterValue(MathOperations::TERM_PARAM__FIELDS, []);
@@ -46,6 +46,8 @@ class MathOperationTotal extends MathOperationStrategy
             );
         }
 
-        return $total;
+        return new ResultNumeric([
+            ResultNumeric::FIELD__NUMBER => $total
+        ]);
     }
 }