+++ src/interfaces/terms/jira/results/ICalculationResult.php
@@ -1,0 +1,20 @@
+<?php
+namespace extas\interfaces\terms\jira\results;
+
+use extas\interfaces\IItem;
+
+/**
+ * Interface ICalculationResult
+ *
+ * @package extas\interfaces\terms\jira
+ * @author jeyroik <jeyroik@gmail.com>
+ */
+interface ICalculationResult extends IItem
+{
+    public const SUBJECT = 'extas.term.jira.calculation.result';
+
+    /**
+     * @return mixed
+     */
+    public function export();
+}